rename globals to constants (they are not in global scope)
